added files and made more compact classes
