+++ Encryption-Decryption/task/src/encryptdecrypt/AlgorithmMethod.java
@@ -1,0 +1,5 @@
+package encryptdecrypt;
+
+public interface AlgorithmMethod {
+    String implement(String input, int key, String alg);
+}
